feat: KB Index architecture — Phase 0a + 0b
Phase 0a: Float priority (u8 → f32)
- OrderedPriority newtype with Ord via f32::total_cmp()
- Migrate OperationConfig, FlushOptions, QueueEvent, BTreeMap
- Persistence: INT64 → DOUBLE for priority column

Phase 0b: Cross-entity KB ingestion
- Schema: {KB}_Index, {KB}_Index_Chunk, {Entity}_SOURCED_{KB} rels
- AggregateProcessor: rebuild _content, chunk per source field, emit
  downstream InsertOp/LinkOp/EmbedOp with priority_override
- _content_offset INT64 on chunks for BM25 highlight translation
- Fix highlight resolution: _content key + offset translation
- title_max_chars in ChunkingConfig (default 256)
- delete/update contentFor-only: propagate to title entities
- link() incremental: detect content→title KB rels, enqueue AggregateOp

362 tests, 0 failures

+++ extension/rag3weaver/src/persistence.rs
@@ -14,7 +14,7 @@ use crate::queue::OperationItem;
 pub struct PersistedOp {
     pub uuid: String,
     pub op_type: String,
-    pub priority: u8,
+    pub priority: f32,
     pub state: String,
     pub temp_uuid: Option<String>,
     pub entity_name: Option<String>,
